Version finale

+++ src/pages/parameters.php
@@ -17,25 +17,36 @@
             include '../requests/Model.php';
 
 
-            if (isset($_POST['pass_old'])) {
-                $old_pass = $_POST['pass_old'];
-                $new_pass = $_POST['new_pass'];
-                $conf_new_pass = $_POST['new_pass_confirmation'];
+            if (isset($_POST['submit'])) {
+                if(isset($_POST['pass_old']) && isset($_POST['new_pass']) && isset($_POST['new_pass_confirmation'])
+                && $_POST['new_pass'] != '' && $_POST['new_pass_confirmation'] != '') {
 
-                $stm = Model::getPassword($_SESSION['mail'])['mot_de_passe'];
-                echo $stm;
-                echo $old_pass;
-                if (strcmp($stm, $old_pass) == 0) {
-                    if (strcmp($new_pass, $conf_new_pass) != 0) {
-                         $msg = "Les mots de passe ne sont pas identiques";
+                    $old_pass = $_POST['pass_old'];
+                    $new_pass = $_POST['new_pass'];
+                    $conf_new_pass = $_POST['new_pass_confirmation'];
+
+                    $stm = Model::getPassword($_SESSION['mail'])['mot_de_passe'];
+                    echo $stm;
+                    echo $old_pass;
+                    if (strcmp($stm, $old_pass) == 0) {
+                        if (strcmp($new_pass, $conf_new_pass) != 0) {
+                            $msg = "Les mots de passe ne sont pas identiques";
+                        } else {
+                            Model::setPassword($_SESSION['mail'], $new_pass);
+                            echo '<script type="text/javascript"> document.location.replace(\'index.php?valid=true&msg=0\');</script>';
+                        }
                     } else {
-                        Model::setPassword($_SESSION['mail'], $new_pass);
-                        echo '<script type="text/javascript"> document.location.replace(\'index.php?valid=true&msg=0\');</script>';
+                        $msg = "Votre mot de passe actuel n'est pas correct";
                     }
                 } else {
-                    $msg = "Votre mot de passe actuel n'est pas correct";
+                    $msg = "Veuillez saisir les champs demandés";
                 }
 
+            }
+
+            if(isset($_POST['delete'])) {
+                Model::deleteProfil($_SESSION['mail']);
+                echo '<script type="text/javascript"> document.location.replace(\'logout.php\');</script>';
             }
 
             if(isset($_GET['error'])) {            
@@ -54,15 +65,14 @@
             <article id="m-parameters-article">
                 <h1>Paramètres</h1>
                 <form method="post">
-                    <input type="file">
-                    <input type="password" name="pass_old" placeholder="Ancien mot de passe" required>
+                    <input type="password" name="pass_old" placeholder="Ancien mot de passe">
 
-                    <input type="password" name="new_pass" placeholder="Nouveau mot de passe" required>
+                    <input type="password" name="new_pass" placeholder="Nouveau mot de passe">
 
-                    <input type="password" name="new_pass_confirmation" placeholder="Confirmation mot de passe"
-                           required>
+                    <input type="password" name="new_pass_confirmation" placeholder="Confirmation mot de passe">
 
                     <input id="parameters-button" type="submit" value="Mettre à jour profil" name="submit">
+                    <input id="user-delete" type="submit" value="Supprimer son compte" name="delete">
                 </form>
                 <?php
                     if(!empty($msg)) {
